de-duplicate format_bytes, quant_dtype parsing
- Add util::format_bytes(u64) shared helper; remove private format_bytes
  from bench.rs and human_size from rm.rs
- Add ServeArgs::resolve_quant_dtype() method; replace the identical 5-line
  quantize-parse block in server.rs, run.rs, and bench.rs with a single call
- Add engine::load_engine() / EngineContext; replace ~40-50 lines of repeated
  download → config → model → Engine → paged-KV boilerplate in all three
  command handlers with a single load_engine(&args) call
